Adding support for
#289
#373
