+++ package.json
@@ -157,6 +157,16 @@
                     "type": "boolean",
                     "default": false,
                     "description": "Ask gist name upon creating. Helps you to identify the gist if you have multiple gists."
+                },
+                "sync.removeExtensions": {
+                    "type": "boolean",
+                    "default": true,
+                    "description": "Set it to false if you dont want to remove extensions while downloading."
+                },
+                "sync.syncExtensions": {
+                    "type": "boolean",
+                    "default": true,
+                    "description": "Set it to false if you dont want to upload / download the extensions."
                 }
             },
             "title": "Code Settings Sync Configuration Settings"
@@ -169,6 +179,8 @@
         "test": "node ./node_modules/vscode/bin/test"
     },
     "devDependencies": {
+        "@types/mocha": "^2.2.48",
+        "mocha": "^5.0.1",
         "typescript": "^2.2.2",
         "vscode": "^1.1.0",
         "@types/node": "^9.4.1"
