+++ src/githubService.ts
@@ -1,7 +1,7 @@
 "use strict";
 
 import * as envir from './environmentPath';
-import * as fileManager from './fileManager';
+import {File, FileService} from './service/fileService';
 import * as vscode from 'vscode';
 import { Environment } from './environmentPath';
 
@@ -80,7 +80,7 @@ export class GitHubService {
         }
     }
 
-    public AddFile(list: Array<fileManager.File>, GIST_JSON_b: any) {
+    public AddFile(list: Array<File>, GIST_JSON_b: any) {
         for (var i = 0; i < list.length; i++) {
             var file = list[i];
             if (file.content != '') {
@@ -123,7 +123,7 @@ export class GitHubService {
         });
     }
 
-    public async CreateAnonymousGist(publicGist: boolean, files: Array<fileManager.File>, gistDesciption: string): Promise<any> {
+    public async CreateAnonymousGist(publicGist: boolean, files: Array<File>, gistDesciption: string): Promise<any> {
         var me = this;
         if (publicGist) {
             me.GIST_JSON_EMPTY.public = true;
@@ -170,7 +170,7 @@ export class GitHubService {
         });
     }
 
-    public UpdateGIST(gistObject: any, files: Array<fileManager.File>): any {
+    public UpdateGIST(gistObject: any, files: Array<File>): any {
 
         var me = this;
         var allFiles: string[] = Object.keys(gistObject.data.files);
