+++ test/extension.test.ts
@@ -5,18 +5,16 @@
 
 // The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
-
+import 'mocha';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
 import * as myExtension from '../src/extension';
+import { Environment } from '../src/environmentPath';
+import Commons from '../src/commons';
 
-// Defines a Mocha test suite to group tests of similar kind together
-// suite("Extension Tests", () => {
-
-// 	// Defines a Mocha unit test
-// 	test("Something 1", () => {
-// 		assert.equal(-1, [1, 2, 3].indexOf(5));
-// 		assert.equal(-1, [1, 2, 3].indexOf(0));
-// 	});
-// });
+describe('Extensions', function(){
+	describe('Delete Extensions', function(){
+		
+	});
+  });
